requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
faiss-cpu==1.12.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==1.10.13
openai==1.14.1
markdown==3.5.2
requests==2.31.0
streamlit